requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pymysql==1.1.0
Flask==2.3.3
beautifulsoup4==4.12.2
gunicorn==21.2.0
